+++ .gitignore
@@ -109,3 +109,4 @@ dist
 # Folder for unit test output
 /.tmp
 *.bak
+tmp.ts
