+++ tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "lib": ["es2019"],
     "module": "commonjs",
     "esModuleInterop": true,
     "target": "es2019",
@@ -11,6 +12,12 @@
     "typeRoots": [
       "node_modules/@types",
       "types"
+    ],
+    "types": [
+      "istanbul-lib-coverage",
+      "istanbul-lib-report",
+      "apex-parser",
+      "jest-stare"
     ]
   },
   "exclude": ["dist", "test", "gulpfile.ts"]
